Work with slave nodes.
The code was assuming that everything happened on the master. It was
getting the workspace directory, but unfortunately it was the workspace
directory on the slave. When the master tried to open files there, the
local path was wrong.

I think this fixes the problem but it looks ugly and is probably the wrong
solution.